Refactor mono config to non-mono config
* Change terraform to not grant cleanup/apiserver access to the CSRF secret.

Fixes #71
