+++ cmd/cleanup/main.go
@@ -38,7 +38,7 @@ import (
 
 func main() {
 	ctx := context.Background()
-	config, err := config.New(ctx)
+	config, err := config.NewCleanupConfig(ctx)
 	if err != nil {
 		log.Fatalf("config error: %v", err)
 	}
