+++ terraform/service_apiserver.tf
@@ -44,13 +44,6 @@ resource "google_secret_manager_secret_iam_member" "apiserver-db" {
   member    = "serviceAccount:${google_service_account.apiserver.email}"
 }
 
-resource "google_secret_manager_secret_iam_member" "apiserver-csrf" {
-  provider  = google-beta
-  secret_id = google_secret_manager_secret.csrf-token.id
-  role      = "roles/secretmanager.secretAccessor"
-  member    = "serviceAccount:${google_service_account.apiserver.email}"
-}
-
 resource "google_cloud_run_service" "apiserver" {
   name     = "apiserver"
   location = var.region
